proposal/requirements.txt
exa-py==1.0.8
httpx==0.27.0
psycopg[binary]==3.2.7
python-dotenv==1.0.1
supabase>=2.3.0,<3
